Fixed RobTarget Instructions
The old version was using the GetPosition() function of PCSDK which returned the extax field as all zeros. This caused the robot arms to not move properly. New version uses GetRobTarget() which returns proper robtarget. Also adjusted code so that only the left or right arm is queried based on which editor the user is creating the variable from (left/right).