New Location Added to Dropdown
Changed the dropdownCreate() function inside of /blockly/core/field_variable.js so that when a new block is added to the workspace the variable dropdown will only offer the user a "New Location" option for that particular block. Other dropdowns are unaffected. Also changed /blockly/msg/messages.js so that the dropdown option read as "Location" and not variable. Then recompiled blockly via ./build.py so that the en.js file would be changed.