Modified Blockly and Rebuilt
In order to create a variable dropdown targeted to our purpose, I changed the rename and delete variable options inside of "blockly/msg/messages.js". I then added a piece of code at the end of the dropdowncreate() function inside of "blockly/core/field_variables.js" which creates the dropdown based on whether or not the block is new to the workspace and has not had a location associated with it.