Fixed Open/Close Gripper Functionality
Placed a tool definition for the servo gripper inside of index.html. Referenced this gripper when writing RobTargets and changed the open/close grip instructions to g_JogIn/g_JogOut. These instructions open/close the grippers to their maximum possible positions or until the gripper encounters an object when closing.