"Wait For Each Other" Sync Block Mostly Works
The basic sync block works for the most part. Sync variables are compiled into code and added at the top of the module. Stopping the program when sync blocks are involved still needs to be worked out, because program execution status will not leave run mode if one arm is waiting on another when the program is stopped from duplo app.