Implemented Scrolling and Adapted drag across for it
Added default blockly scrollbars to workspaces. Overrode two functions so that when one workspace is scrolled the scroll bars of the other workspace are adjusted accordingly. Also had to add in a scrolled workspace offset for the drag across function when a block is being moved from one workspace to another. I'm not completely happy with the outcome, but it works fairly well and short of updating our blockly version, (they've made a lot of changes since and that would be a pint), it should probably do.