Move webview postMessage into confirm modal click handler
Robtargets were not being captured properly because the request arm position message had to be in the modal confirm click handler since it is asynchronous. Also added a handler for the cancel button on the modal and set the modal to static so that it cannot be hidden in any other way than the confirm or cancel buttons.