Removed WPF Textbox and added reset robot button to JS webpage
Still need to synchronize the buttons with some sort of a "mutex lock" so that multiple write requests to the robot are not allowed until the current one is halted by error, program end, or the stop button.

Also added save image in place of text box which could ultimately serve as a pop up modal for saving the blockly workspaces? Not quite sure how this would work yet.